+++ src/wrath_header/decrypt.rs
@@ -10,6 +10,7 @@ use std::io::Read;
 /// Decryption part of a [`ServerCrypto`](crate::wrath_header::ServerCrypto).
 ///
 /// Intended to be kept with the reader half of a connection.
+#[derive(Debug)]
 pub struct ServerDecrypterHalf {
     decrypt: InnerCrypto,
 }
@@ -64,6 +65,7 @@ impl ServerDecrypterHalf {
 /// Decryption part of a [`ClientCrypto`](crate::wrath_header::ClientCrypto).
 ///
 /// Intended to be kept with the reader half of a connection.
+#[derive(Debug)]
 pub struct ClientDecrypterHalf {
     decrypt: InnerCrypto,
 }